Add ratings to leaderboards
